fix(v9.12.0): Pegelexplosion + L/R-Zeitversatz + Goals-Kalibrierung
Kernfixes:
- phase_12: §2.51 M/S-Domain-Fix — verhindert PSOLA L/R-Zeitversatz (8777 Samples, 182ms).
  ROOT CAUSE war unabhängiges PSOLA pro Kanal mit verschiedenen pitch/period_samps.
  FIX: Mid=(L+R)/2 → PSOLA, Side=(L-R)/2 → Phase-Vocoder, identisches Timing.
- phase_12/30: §2.45a-I Gated-RMS statt globalem RMS (kein Pegelexplosion im Tail)
- phase_29: sosfiltfilt statt sosfilt in Bandpass-Chain (inter-Band Zeitversatz→Pegelexplosion)
- phase_42: sosfiltfilt in _reduce_harshness/_apply_deessing/_control_breath
- phase_36/18/11/10/13/15/17/22/33/35/37/39/44/45: sosfiltfilt (split+recombine)
- phase_23: §2.30c POCS-Projektion vor PGHI, ADMM wall-time adaptiv, SyntaxError behoben
- micro_dynamics: §2.30b Post-SG-Quiet-Zone-Clamp + Frisson Zwei-Stufen-Invariante (Pre+Post-SG)
- emotional_arc: §2.30b Per-Sample-Guard -36 dBFS in correct_arc()
- musical_goals: TonalCenter Soft-Floor, GrooveMetric bidirektionaler DTW-Guard,
  AuthentizitaetMetric formant_threshold max(1200, ref*1.5), waerme Proxy / 4.0
- §0d carrier_chain_recovery_ratio Pflichtfeld in UV3 metadata
- §C10 SongGoalFeedbackStore Bayesian-EMA-Blend
- §Frisson FrissonCandidateDetector → MDEM frisson_zones Pipeline
- §2.30c WaveformPlausibilityGuard nach correct_arc in UV3
- PMGG delta-adaptiver Blend (§09.2), absoluter Headroom-Scalar (§2.54)
- adaptive_thresholds Material-Ceiling Dict
- dsp/: docstring-after-code Fix, F821-Fix, shebang-Berechtigungen

+++ backend/core/phases/output_guard.py
@@ -19,7 +19,7 @@ class OutputGuardDecision:
 
 def rms(audio: np.ndarray) -> float:
     """Return RMS for mono/stereo audio."""
-    x = np.nan_to_num(audio.astype(np.float64), nan=0.0, posinf=0.0, neginf=0.0)
+    x = np.nan_to_num(audio.astype(np.float32), nan=0.0, posinf=0.0, neginf=0.0)
     return float(np.sqrt(np.mean(x**2) + 1e-12))
 
 
@@ -29,10 +29,10 @@ def side_rms(audio: np.ndarray) -> float:
         return 0.0
     # (N, 2) samples-first
     if audio.shape[1] == 2 and audio.shape[0] > 2:
-        side = 0.5 * (audio[:, 0].astype(np.float64) - audio[:, 1].astype(np.float64))
+        side = 0.5 * (audio[:, 0].astype(np.float32) - audio[:, 1].astype(np.float32))
     # (2, N) channels-first
     elif audio.shape[0] == 2 and audio.shape[1] > 2:
-        side = 0.5 * (audio[0].astype(np.float64) - audio[1].astype(np.float64))
+        side = 0.5 * (audio[0].astype(np.float32) - audio[1].astype(np.float32))
     else:
         return 0.0
     return float(np.sqrt(np.mean(side**2) + 1e-12))
